Cadeau les boys

+++ src/ch/hearc/business/Cook.java
@@ -3,7 +3,7 @@ package ch.hearc.business;
 // Arc pour la cuisine
 public class Cook extends Edge {
 
-    public Cook(String name, Receipt ReceiptDest, double metric){
+    public Cook(String name, Receipt ReceiptDest, int metric){
         super(name, ReceiptDest, metric);
     }
 }
