+++ src/ch/hearc/business/Edge.java
@@ -4,13 +4,13 @@ import java.io.Serializable;
 
 public class Edge implements Serializable {
     private String name;
-    private double metric;
+    private int metric;
     private Node dest;
 
     /**
      * Constructeur
      */
-    public Edge(String name, Node dest, double metric) {
+    public Edge(String name, Node dest, int metric) {
         this.name = name;
         this.metric = metric;
         this.dest = dest;
@@ -32,7 +32,7 @@ public class Edge implements Serializable {
         return name;
     }
 
-    public double getMetric() {
+    public int getMetric() {
         return metric;
     }
