+++ src/ch/hearc/business/IsFriend.java
@@ -2,7 +2,7 @@ package ch.hearc.business;
 
 public class IsFriend extends Edge {
 
-    public IsFriend(String name, Person personDest, double metric){
+    public IsFriend(String name, Person personDest, int metric){
         super(name, personDest, metric);
     }
 }
