+++ src/ch/hearc/business/Watch.java
@@ -2,7 +2,7 @@ package ch.hearc.business;
 
 public class Watch extends Edge {
 
-    public Watch(String name, WebSite WebSiteDest, double metric){
+    public Watch(String name, WebSite WebSiteDest, int metric){
         super(name, WebSiteDest, metric);
     }
 }
